Refactor addMines and getBoard functions
