+++ src/game/index.ts
@@ -1,0 +1,6 @@
+import { getBoard } from "./board/getBoard.js";
+import { addMines } from "./mines/addMines.js";
+
+const board = getBoard(5);
+
+addMines(board);
