--- src/ui/board.ts
@@ -1,32 +1,0 @@
-import { Cell, Board } from "./types.js";
-import { addMines } from "./mines.js";
-
-export const getGameBoard = (
-  numberA: number = 5,
-  numberB: number = 5
-): Board => {
-  const board: Board = [];
-  const rowsTotal = numberA;
-  const columnsTotal = numberB;
-
-  if (numberA !== numberB) {
-    throw new Error("There must be the same number of rows and columns");
-  }
-
-  for (let rowNumber = 0; rowNumber < numberA; rowNumber++) {
-    board[rowNumber] = [];
-    for (let columnNumber = 0; columnNumber < numberB; columnNumber++) {
-      const boardCell: Cell = {
-        hasMine: false,
-        adjacentMinesTotal: 0,
-        isOpen: false,
-      };
-      board[rowNumber][columnNumber] = boardCell;
-    }
-  }
-  return board;
-};
-
-let gameBoard = getGameBoard();
-
-addMines(gameBoard);
