--- src/ui/data.ts
@@ -1,7 +1,0 @@
-import { Cell } from "./types.js";
-
-export const boardCell: Cell = {
-  hasMine: false,
-  adjacentMinesTotal: 0,
-  isOpen: false,
-};
